# Back-end/User-API/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API Fitisly",
    "description": "API for the fitlisly app",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "ec2-3-14-244-98.us-east-2.compute.amazonaws.com",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "definitions": {
...
    "exercise": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "repeat_number": {
          "type": "number"
        },
        "rest_time": {
          "type": "number",
          "format": "minute:second"
        },
        "exerciseImage": {
          "type": "number"
        }
      }
    },
...
  },
  "paths": {
...
    "/api/exercises": {
      "parameters": [
        {
          "name": "Authorization",
          "in": "header",
          "description": "Authorization header",
          "required": true,
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Exercises"
        ],
        "summary": "Get all the exercises of the database",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/exercise"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Exercises"
        ],
        "summary": "Create a new exercise in the database",
        "produces": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "exerciseImage",
            "type": "file",
            "description": "image of the exercise"
          },
          {
            "in": "formData",
            "name": "name",
            "type": "string",
            "description": "name of the exercise"
          },
          {
            "in": "formData",
            "name": "description",
            "type": "string",
            "description": "description of the exercise"
          },
          {
            "in": "formData",
            "name": "repeat_number",
            "type": "number",
            "description": "number of repetition of the exercise"
          },
          {
            "in": "formData",
            "name": "rest_time",
            "type": "number",
            "description": "the rest time required after the exercise"
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/exercise"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/exercises/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of the exercise that we want to retrieve",
          "type": "integer"
        },
        {
          "name": "Authorization",
          "in": "header",
          "description": "Authorization header",
          "required": true,
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Exercises"
        ],
        "summary": "Get exercise with the given ID",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/exercise"
            }
          },
          "404": {
            "description": "Exercise not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": ["Exercises"],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "exerciseImage",
            "type": "file",
            "description": "image of the exercise"
          },
          {
            "in": "formData",
            "name": "name",
            "type": "string",
            "description": "name of the exercise"
          },
          {
            "in": "formData",
            "name": "description",
            "type": "string",
            "description": "description of the exercise"
          },
          {
            "in": "formData",
            "name": "repeat_number",
            "type": "number",
            "description": "number of repetition of the exercise"
          },
          {
            "in": "formData",
            "name": "rest_time",
            "type": "number",
            "description": "the rest time required after the exercise"
          }
        ],
        "summary": "Update the exercise with the given ID",
        "responses": {
          "200": {
            "description": "Exercise updated sucessfully",
            "schema": {
              "$ref": "#/definitions/exercise"
            }
          },
          "404": {
            "description": "Exercise not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": ["Exercises"],
        "summary": "Delete exercise with the given ID",
        "responses": {
          "200": {
            "description": "Exercise {id} was deleted sucessfully."
          },
          "404": {
            "description": "Exercise not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}